+++ CHANGELOG.md
@@ -10,6 +10,9 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Support for pre-wash transformations
+- Lock settings that no longer affect the result once other actions have completed successfully,
+  especially to prevent trying to ingest a given Preservica SIP file to multiple environments (which
+  fails when environments are different tenants on the same cloud instance)
 
 
 ## 1.0.0 - 2021-04-01